update api with userId

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getCurrentUserServer } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import * as yup from "yup";
@@ -33,11 +34,18 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: Request) {
+  const user = await getCurrentUserServer();
+  if (!user) {
+    return NextResponse.json("no autorizado", { status: 401 });
+  }
+
   try {
     const { complete, description } = await postSchema.validate(
       await request.json()
     );
-    const todo = await prisma.todo.create({ data: { complete, description } });
+    const todo = await prisma.todo.create({
+      data: { complete, description, userId: user.id },
+    });
 
     return NextResponse.json(todo);
   } catch (error) {
@@ -50,9 +58,14 @@ const deleteSchema = yup.object({
 });
 
 export async function DELETE(request: Request) {
+  const user = await getCurrentUserServer();
+  if (!user) {
+    return NextResponse.json("no autorizado", { status: 401 });
+  }
+
   try {
     const response = await prisma.todo.deleteMany({
-      where: { complete: true },
+      where: { complete: true, userId: user.id },
     });
 
     return NextResponse.json(response);
